[MR] Separated fp-related functions from the rest of the dsl
