+++ code_gen/cpp_gen.rb
@@ -1,29 +1,14 @@
 require 'Utility/helper_cpp'
 require 'Utility/fp_helper_cpp'
+require 'code_gen/fp_cpp_gen'
 # frozen_string_literal: true
 
 # Semantics Generator: Converts IR to C++ code
 module CodeGen
   class CppGenerator
-    attr_reader :emitter, :mapping
+    include FPEmitters
 
-    # PROPOSAL:
-    # Add RM (rounding modes) map that converts
-    # RV rounding modes into softfloat rounding modes.
-    # They actually are exactly the same, but we can generalize this idea later.
-    # Maybe should move it to other module
-    # RISC-V RM to SoftFloat RM mapping
-    # RNE=0 -> softfloat_round_near_even(0), RTZ=1 -> softfloat_round_minMag(1),
-    # RDN=2 -> softfloat_round_min(2), RUP=3 -> softfloat_round_max(3),
-    # RMM=4 -> softfloat_round_near_maxMag(4), DYN=7 -> read from fcsr
-    SOFTFLOAT_RM_MAP = {
-      0 => 'softfloat_round_near_even',
-      1 => 'softfloat_round_minMag',
-      2 => 'softfloat_round_min',
-      3 => 'softfloat_round_max',
-      4 => 'softfloat_round_near_maxMag',
-      7 => nil # DYN - handled specially
-    }.freeze
+    attr_reader :emitter, :mapping
 
     def initialize(emitter, mapping = {})
       @emitter = emitter
@@ -39,278 +24,6 @@ module CodeGen
       src2 = src2.nil? ? operation[:oprnds][2][:value] : src2
 
       emitter.emit_line("#{dst} = #{src1} #{op_str} #{src2};")
-    end
-
-    # PROPOSAL:
-    # Add FP emitter helpers
-    # Emit code to set SoftFloat rounding mode before FP operation
-    # For most instruction in softfloat, rounding mode is defined
-    # by global variable, so we have to change it before we execute.
-    # We also store it into a temporal variable, so we can restore it later.
-    def emit_rm_setup(rm, tmp_var = '_rm_save')
-      return unless rm # nil means no rm handling needed
-
-      # Currently not supported
-      if rm == 7 # DYN mode - read from fcsr at runtime
-        @emitter.emit_line('assert(0 && "CSR is currently not supported\n");')
-        @emitter.emit_line("uint_fast8_t #{tmp_var} = softfloat_roundingMode;")
-        # @emitter.emit_line('softfloat_roundingMode = cpu.getFCSR_RM();')
-      else
-        rm_const = SOFTFLOAT_RM_MAP[rm]
-        @emitter.emit_line("uint_fast8_t #{tmp_var} = softfloat_roundingMode;")
-        @emitter.emit_line("softfloat_roundingMode = #{rm_const};")
-      end
-    end
-
-    # Emit code to restore SoftFloat rounding mode after FP operation
-    def emit_rm_restore(tmp_var = '_rm_save')
-      @emitter.emit_line("softfloat_roundingMode = #{tmp_var};")
-    end
-
-    # Emits binary fp operation.
-    # This is different from RV32I binop, because:
-    # 1. typing is different (check Utility/fp_helper_cpp.rb)
-    # 2. Rounding mode has to be changed
-    def emit_fp_binary(opname, operation, dst_type:, src_types:, rm: nil)
-      dst  = map_operand(operation[:oprnds][0])
-      src1 = map_operand(operation[:oprnds][1])
-      src2 = map_operand(operation[:oprnds][2])
-
-      t1 = Utility.gen_typed_tmp(src1, src_types[0])
-      t2 = Utility.gen_typed_tmp(src2, src_types[1])
-      tr = Utility.gen_typed_tmp(dst,  dst_type)
-
-      if Utility::FP_INFO[src_types[0]]
-        info = Utility::FP_INFO[src_types[0]]
-        @emitter.emit_line("#{info[:c_type]} #{t1} = { #{info[:unpack]}(#{src1}) };")
-      else
-        ctype = Utility::INT_INFO[src_types[0]]
-        @emitter.emit_line("#{ctype} #{t1} = (#{ctype})(#{src1});")
-      end
-
-      if Utility::FP_INFO[src_types[1]]
-        info = Utility::FP_INFO[src_types[1]]
-        @emitter.emit_line("#{info[:c_type]} #{t2} = { #{info[:unpack]}(#{src2}) };")
-      else
-        ctype = Utility::INT_INFO[src_types[1]]
-        @emitter.emit_line("#{ctype} #{t2} = (#{ctype})(#{src2});")
-      end
-
-      # Set rounding mode before operation if rm is specified
-      emit_rm_setup(rm) if rm
-
-      if Utility::FP_INFO[dst_type]
-        ctype = Utility::FP_INFO[dst_type][:c_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{t1}, #{t2});")
-        pack = Utility::FP_INFO[dst_type][:pack]
-        @emitter.emit_line("#{dst} = #{pack % tr};")
-      else
-        ctype = Utility::INT_INFO[dst_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{t1}, #{t2});")
-        @emitter.emit_line("#{dst} = (uint64_t)#{tr};")
-      end
-
-      # Restore rounding mode after operation if rm was set
-      emit_rm_restore if rm
-    end
-
-    # float -> int conversions need to be handled separetely, because
-    # they break general pattern of softfloat instructions by demanding
-    # rounding mode as an argument (they ignore global constant that all other
-    # fp functions use for some unknown reason).
-    def emit_fp_to_int_conv(opname, operation, dst_type:, src_type:, rounding_mode:, exact: true)
-      dst = map_operand(operation[:oprnds][0])
-      src = map_operand(operation[:oprnds][1])
-
-      ts = Utility.gen_typed_tmp(src, src_type)
-      tr = Utility.gen_typed_tmp(dst, dst_type)
-
-      if Utility::FP_INFO[src_type]
-        info = Utility::FP_INFO[src_type]
-        @emitter.emit_line("#{info[:c_type]} #{ts} = { #{info[:unpack]}(#{src}) };")
-      else
-        ctype = Utility::INT_INFO[src_type]
-        @emitter.emit_line("#{ctype} #{ts} = (#{ctype})(#{src});")
-      end
-
-      exact_arg = exact ? '1' : '0'
-      @emitter.emit_line("#{Utility::INT_INFO[dst_type]} #{tr} = #{opname}(#{ts}, #{rounding_mode}, #{exact_arg});")
-      @emitter.emit_line("#{dst} = (uint64_t)#{tr};")
-    end
-
-    # There is no min/max/neg functions in softfloat,
-    # so we have to implement them ourselves
-    def emit_fp_min_max(opname, operation, type:)
-      dst, src1, src2 = map_n_operands(operation, 3)
-      t1 = Utility.gen_typed_tmp(src1, type)
-      t2 = Utility.gen_typed_tmp(src2, type)
-      tr = Utility.gen_typed_tmp(dst, type)
-      info = Utility::FP_INFO[type]
-      prefix = type == :f32 ? 'f32' : 'f64'
-      sign_bit = type == :f32 ? 31 : 63
-
-      @emitter.emit_line("#{info[:c_type]} #{t1} = { #{info[:unpack]}(#{src1}) };")
-      @emitter.emit_line("#{info[:c_type]} #{t2} = { #{info[:unpack]}(#{src2}) };")
-      @emitter.emit_line("#{info[:c_type]} #{tr};")
-
-      # Bitwise NaN detection. Couldn't find softfloat's one, also this would faster with JIT
-      # in comparison to function call.
-      if type == :f32
-        @emitter.emit_line("bool _a_nan = ((#{t1}.v >> 23) & 0xFF) == 0xFF && (#{t1}.v & 0x7FFFFF) != 0;")
-        @emitter.emit_line("bool _b_nan = ((#{t2}.v >> 23) & 0xFF) == 0xFF && (#{t2}.v & 0x7FFFFF) != 0;")
-      else
-        @emitter.emit_line("bool _a_nan = ((#{t1}.v >> 52) & 0x7FF) == 0x7FF && (#{t1}.v & 0xFFFFFFFFFFFFF) != 0;")
-        @emitter.emit_line("bool _b_nan = ((#{t2}.v >> 52) & 0x7FF) == 0x7FF && (#{t2}.v & 0xFFFFFFFFFFFFF) != 0;")
-      end
-
-      is_min = opname.include?('min')
-
-      @emitter.emit_line('{')
-      # return number if one of them is NaN, if both are NaN, return second
-      @emitter.emit_line("  if (_a_nan && _b_nan) #{tr} = #{t2};")
-      @emitter.emit_line("  else if (_a_nan) #{tr} = #{t2};")
-      @emitter.emit_line("  else if (_b_nan) #{tr} = #{t1};")
-      @emitter.emit_line('  else {')
-      if is_min
-        # smaller or neg zero
-        @emitter.emit_line("    if (#{prefix}_lt(#{t1}, #{t2}) || (#{prefix}_eq(#{t1}, #{t2}) && ((#{t1}.v >> #{sign_bit}) & 1))) #{tr} = #{t1};")
-      else
-        # larger or pos zero
-        @emitter.emit_line("    if (#{prefix}_lt(#{t2}, #{t1}) || (#{prefix}_eq(#{t1}, #{t2}) && ((#{t2}.v >> #{sign_bit}) & 1))) #{tr} = #{t1};")
-      end
-      @emitter.emit_line("    else #{tr} = #{t2};")
-      @emitter.emit_line('  }')
-      @emitter.emit_line('}')
-
-      pack = info[:pack]
-      @emitter.emit_line("#{dst} = #{pack % tr};")
-    end
-
-    def gen_fp_neg(tmp_name, type)
-      case type
-      when :f32 then "#{tmp_name}.v ^= 0x80000000U;"
-      when :f64 then "#{tmp_name}.v ^= 0x8000000000000000ULL;"
-      end
-    end
-
-    # same as above
-    def emit_fp_unary(opname, operation, dst_type:, src_type:, rm: nil)
-      dst = map_operand(operation[:oprnds][0])
-      src = map_operand(operation[:oprnds][1])
-
-      ts = Utility.gen_typed_tmp(src, src_type)
-      tr = Utility.gen_typed_tmp(dst, dst_type)
-
-      if Utility::FP_INFO[src_type]
-        info = Utility::FP_INFO[src_type]
-        @emitter.emit_line("#{info[:c_type]} #{ts} = { #{info[:unpack]}(#{src}) };")
-      else
-        ctype = Utility::INT_INFO[src_type]
-        @emitter.emit_line("#{ctype} #{ts} = (#{ctype})(#{src});")
-      end
-
-      # Set rounding mode before operation if rm is specified
-      emit_rm_setup(rm) if rm
-
-      if Utility::FP_INFO[dst_type]
-        ctype = Utility::FP_INFO[dst_type][:c_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{ts});")
-        pack = Utility::FP_INFO[dst_type][:pack]
-        @emitter.emit_line("#{dst} = #{pack % tr};")
-      else
-        ctype = Utility::INT_INFO[dst_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{ts});")
-        @emitter.emit_line("#{dst} = (uint64_t)#{tr};")
-      end
-
-      # Restore rounding mode after operation if rm was set
-      emit_rm_restore if rm
-    end
-
-    # RV64F has FMA instructions that have 3 sources, so this
-    # new emitter is necessary.
-    def emit_fp_ternary(opname, operation,
-                        dst_type:,
-                        src_types:,
-                        negate_src: [],
-                        rm: nil)
-      dst  = map_operand(operation[:oprnds][0])
-      src1 = map_operand(operation[:oprnds][1])
-      src2 = map_operand(operation[:oprnds][2])
-      src3 = map_operand(operation[:oprnds][3])
-
-      srcs  = [src1, src2, src3]
-      svars = []
-
-      srcs.each_with_index do |src, i|
-        ty = src_types[i]
-        t  = Utility.gen_typed_tmp(src, ty)
-        svars << t
-
-        if Utility::FP_INFO[ty]
-          info = Utility::FP_INFO[ty]
-          @emitter.emit_line("#{info[:c_type]} #{t} = { #{info[:unpack]}(#{src}) };")
-          if negate_src.include?(i)
-            # TODO: change for gen_fp_neg
-            negation = gen_fp_neg(t, ty)
-            @emitter.emit_line(negation)
-            # negf = ty == :f32 ? 'f32_neg' : 'f64_neg'
-            # @emitter.emit_line("#{t} = #{negf}(#{t});")
-          end
-
-        else
-          ctype = Utility::INT_INFO[ty]
-          @emitter.emit_line("#{ctype} #{t} = (#{ctype})(#{src});")
-
-          @emitter.emit_line("#{t} = -#{t};") if negate_src.include?(i)
-        end
-      end
-
-      tr = Utility.gen_typed_tmp(dst, dst_type)
-
-      # Set rounding mode before operation if rm is specified
-      emit_rm_setup(rm) if rm
-
-      if Utility::FP_INFO[dst_type]
-        ctype = Utility::FP_INFO[dst_type][:c_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{svars.join(', ')});")
-        pack = Utility::FP_INFO[dst_type][:pack]
-        @emitter.emit_line("#{dst} = #{pack % tr};")
-      else
-        ctype = Utility::INT_INFO[dst_type]
-        @emitter.emit_line("#{ctype} #{tr} = #{opname}(#{svars.join(', ')});")
-        @emitter.emit_line("#{dst} = (uint64_t)#{tr};")
-      end
-
-      # Restore rounding mode after operation if rm was set
-      emit_rm_restore if rm
-    end
-
-    # No softfloat lib function, so hand-written again
-    def emit_sign_inject(operation, width:, mode:)
-      dst, src1, src2 = map_n_operands(operation, 3)
-
-      mag_mask, sign_mask =
-        if width == 32
-          [Utility::F32_MAG_MASK, Utility::F32_SIGN_MASK]
-        else
-          [Utility::F64_MAG_MASK, Utility::F64_SIGN_MASK]
-        end
-
-      sign_expr =
-        case mode
-        when :copy   then "(#{src2} & #{sign_mask})"
-        when :neg    then "(~#{src2} & #{sign_mask})"
-        when :xor    then "(#{src1} ^ (#{src2} & #{sign_mask}))"
-        end
-
-      result = if mode == :xor
-                 sign_expr
-               else
-                 "(#{src1} & #{mag_mask}) | #{sign_expr}"
-               end
-      result = "0xFFFFFFFF00000000ULL | (uint64_t)(#{result})" if width == 32
-      @emitter.emit_line("#{dst} = #{result};")
     end
 
     # Helpers for easier operand mapping in operations emitters
